# docs/openapi-platform.yaml
openapi: 3.0.3
info:
  title: Plantbot Platform API (session face)
  version: "1.0"
  description: |
    The complete browser/session-facing API — everything the Plantbot console
    itself uses, documented for host webapps that embed Plantbot (iframe) or
    drive it on behalf of signed-in users.

    **Auth**: HMAC-signed session cookie `pb_sess`, obtained via
    `POST /api/auth/login` (password) or the OIDC SSO flow
    (`GET /api/auth/oidc/login` → IdP → callback). Roles are
    `viewer < operator < admin`, granted per site (or `*` platform-wide).
    Anonymous requests carry the viewer role unless the deployment sets
    `PB_PUBLIC_VIEW=0`. For cross-site iframes set `PB_COOKIE_SAMESITE=none`
    (HTTPS required).

    **Realtime**: `GET /ws?site=<siteId>` (WebSocket) opens with a `hello`
    snapshot, then pushes frames tagged by a `t` field — the same data plane the
    console renders. Frame types (`t`): `hello`, `tel` (telemetry, 4 Hz),
    `site`, `geo`, `fleet`, `event`, `lifecycle`, `readings`, `command`,
    `missions`, `schedules`, `templates`, `rules`, `eventTypes`, `channels`,
    `maps`.

    The machine-to-machine integration face (`/api/integration/v1`, Bearer
    site key) is a separate spec: `docs/openapi.yaml`, served live at
    `GET /api/integration/v1/openapi.json`.
servers:
  - url: /
...
security:
  - cookieAuth: []
paths:
...
  /api/auth/oidc/login:
    get:
      tags: [auth]
      security: []
      summary: Begin OIDC SSO (Authorization Code + PKCE)
      description: 302 to the configured IdP. Enabled when the deployment sets OIDC_ISSUER + OIDC_CLIENT_ID.
      operationId: oidcLogin
      parameters:
        - { name: next, in: query, description: "in-app path to land on after sign-in (default /)", schema: { type: string } }
      responses:
        "302": { description: redirect to the IdP authorization endpoint }
        "404": { description: SSO not configured }
        "502": { description: IdP discovery failed }
  /api/auth/oidc/callback:
    get:
      tags: [auth]
      security: []
      summary: OIDC redirect URI
      description: Validates state, exchanges the code, verifies the ID token against the IdP JWKS, JIT-provisions the user and issues the pb_sess session.
      operationId: oidcCallback
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses:
        "302": { description: session issued — redirect into the SPA }
        "400": { description: state mismatch }
        "401": { description: token exchange / validation failed }
...
  /api/users:
    get:
      tags: [platform]
      summary: List users (platform admin)
      operationId: listUsers
      responses:
        "200": { description: users, content: { application/json: { schema: { type: object, properties: { users: { type: array, items: { $ref: "#/components/schemas/User" } } } } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [platform]
      summary: Create user (platform admin)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, roles]
              properties:
                username: { type: string }
                displayName: { type: string }
                password: { type: string }
                roles: { $ref: "#/components/schemas/RoleMap" }
      responses:
        "200": { description: created }
        "400": { description: invalid / duplicate username }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pb_sess
      description: HMAC-signed session cookie from /api/auth/login or the OIDC flow
...
  responses:
    Forbidden: { description: requires a higher role on this site }
...
  schemas:
    Role: { type: string, enum: [viewer, operator, admin] }
    RoleMap:
      type: object
      description: "siteId (or '*') → role"
      additionalProperties: { $ref: "#/components/schemas/Role" }
    User:
      type: object
      properties:
        username: { type: string }
        displayName: { type: string }
        roles: { $ref: "#/components/schemas/RoleMap" }
...